This is how to make a div appear above the current one
make a div appear above the current one on a portfolio, or anything else. Just make sure the divs are inside of a container (parent) for position absolute. Make one of the div's in the css display of none, then in javascript, make the onmouse event change the display from none to block or flex.  see code...